Initial swag at 24-bit color support for Windows consoles.
This seems to work well on conemu, but doesn't work at all normally
in the stock Windows 10 console.  However, if you run under env, it
will 'sort of work', meaning once you press enter things seem to work.

This is for experimentation, opt-in only.  To enable this, set the
environment variable TCELL_TRUECOLOR to "enable".

Advice to fix this for console hosts is requested.