Fix off by one cursor addressing, and support for Windows 10 term.
With the above fix, we have surfaced a strange bug in ConEmu, where the
colors scroll, but the content doesn't, when using VT escapes.  When we
detect ConEmu, we default to disabling this mode.

All others are on by default.  It seems to work well with the stock
Terminal.

Opt-Out by setting TCELL_TRUECOLOR=disable.  ConEmu users can forcibly
opt-in by setting TCELL_TRUECOLOR=enable.